# api-specs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Fruition
  description: Fruition PoC 백엔드 API
  title: Fruition API
  version: 0.0.1
security:
- bearerAuth: []
...
paths:
...
  /api/query/runs/{requestId}:
    get:
      description: 비동기 질의의 현재 상태와 완료 결과 또는 오류 정보를 반환합니다.
      operationId: getRun_1
      parameters:
      - description: 비동기 질의 요청 ID
        in: path
        name: requestId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QueryRunStatusResponse"
          description: 상태 조회 성공
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 질의 실행 또는 워크스페이스를 찾을 수 없음
      summary: 질의 실행 상태 조회
      tags:
      - Query Runs
  /api/query/runs/{requestId}/events:
    get:
      description: 비동기 질의의 진행 상황과 최종 결과를 Server-Sent Events로 전달합니다.
      operationId: subscribe
      parameters:
      - description: 비동기 질의 요청 ID
        in: path
        name: requestId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                type: string
          description: SSE 구독 시작
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 질의 실행 또는 워크스페이스를 찾을 수 없음
      summary: 질의 진행 이벤트 구독
      tags:
      - Query Runs
...
  /api/workspaces/{workspace_id}/agent/runs/{run_id}:
    get:
      description: 자율 AgentRun 계획과 실행 상태를 조회합니다.
      operationId: getRun
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - in: path
        name: run_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsonNode"
          description: OK
      summary: AgentRun 조회
      tags:
      - Agent
...
  /api/workspaces/{workspace_id}/wiki/maintenance/runs/{run_id}:
    get:
      description: 실행 중이거나 완료된 Wiki 정합성 검사 결과를 반환합니다.
      operationId: run
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - description: 조회할 검사 실행 ID
        in: path
        name: run_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsonNode"
          description: 결과 조회 성공
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 검사 실행 또는 워크스페이스를 찾을 수 없음
      summary: Wiki 정합성 검사 결과 조회
      tags:
      - Wiki Maintenance
...
components:
  schemas:
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: 분기 기준으로 쓰는 값. message가 아니라 이 값으로 분기한다.
          example: INVALID_REQUEST
        details:
          type: array
          description: 검증 실패(400 INVALID_REQUEST)일 때만 존재한다. 값이 없으면 키 자체가 빠지므로 null 체크가 아니라 존재 여부로 확인한다.
          items:
            $ref: "#/components/schemas/FieldError"
        message:
          type: string
          description: 사용자에게 그대로 보여도 되는 한국어 문구
          example: 요청 형식이 올바르지 않습니다.
    ErrorResponse:
      type: object
      description: 컨트롤러까지 도달한 요청이 실패했을 때 모든 API가 공유하는 에러 응답 형식. 인증 필터가 막는 401은 HttpStatusEntryPoint가 상태 코드만 내보내므로 이 본문이 없다.
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 에러 상세
    EvidenceSnippet:
      type: object
      description: 답변을 뒷받침하는 원문 구절
      properties:
        rank:
          type: integer
          format: int32
          description: 근거 순위(0부터). 낮을수록 강한 근거다.
          example: 0
        source_block_ids:
          type: array
          description: 대표 문서 안에서의 근거 block ID 목록
          items:
            type: string
        source_document_id:
          type: string
          description: 대표 근거 문서 ID. 여러 문서를 참조하면 source_refs를 본다.
          example: doc_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
        source_refs:
          type: array
          description: 문서·block 쌍으로 표현한 전체 근거 위치. 여러 문서를 걸칠 수 있다.
          items:
            $ref: "#/components/schemas/QuerySourceRef"
        text:
          type: string
          description: 근거 구절 원문
    FieldError:
      type: object
      properties:
        field:
          type: string
          description: 문제가 된 요청 필드명
          example: email
        reason:
          type: string
          description: 그 필드의 위반 사유
          example: email은 필수입니다.
...
    GraphContext:
      type: object
      description: 답변에 쓰인 Wiki 그래프의 부분 그래프
      properties:
        edges:
          type: array
          description: 노드 사이의 관계
          items:
            $ref: "#/components/schemas/GraphEdge"
        nodes:
          type: array
          description: 문맥에 포함된 페이지 노드
          items:
            $ref: "#/components/schemas/RelatedPage"
    GraphEdge:
      type: object
      description: 질의 문맥에서의 페이지 간 관계
      properties:
        from_page_id:
          type: string
          description: 시작 페이지 ID
        link_type:
          type: string
          description: 관계 종류
          example: related
        role:
          type: string
          description: 이 관계가 답변에서 맡은 역할
        score:
          type: number
          format: double
          description: 관계 점수
          example: 0.72
        to_page_id:
          type: string
          description: 도착 페이지 ID
...
    JsonNode: {}
...
    MessageSummary:
      type: object
      description: 질의로 저장된 메시지 요약
      properties:
        content:
          type: string
          description: 메시지 본문
        created_at:
          type: string
          format: date-time
          description: 생성 시각(ISO-8601 UTC)
          example: 2026-08-13T04:25:24.371948Z
        id:
          type: string
          description: 메시지 ID
        role:
          type: string
          description: 발화 주체
          enum:
          - user
          - assistant
          example: assistant
        status:
          type: string
          description: 처리 상태
          example: completed
...
    QueryResponse:
      type: object
      description: 질의 결과. 웹 검색 결과 URL과 인증 토큰은 포함되지 않는다.
      properties:
        assistant_message:
          $ref: "#/components/schemas/MessageSummary"
          description: 저장된 답변 메시지
        error_code:
          type: string
          description: 웹 검색이 실패했을 때의 사유
          enum:
          - web_search_unavailable
          - web_search_failed
        evidence_snippets:
          type: array
          description: 답변을 뒷받침하는 원문 구절 목록
          items:
            $ref: "#/components/schemas/EvidenceSnippet"
        graph_context:
          $ref: "#/components/schemas/GraphContext"
          description: 답변에 쓰인 Wiki 그래프 문맥
        related_pages:
          type: array
          description: 답변 근거가 된 Wiki 페이지 목록
          items:
            $ref: "#/components/schemas/RelatedPage"
        result_count:
          type: integer
          format: int32
          description: 근거로 쓴 결과 수
          example: 5
        traversal_paths:
          type: array
          description: 그래프를 따라간 경로. 왜 이 페이지가 근거인지 설명한다.
          items:
            $ref: "#/components/schemas/TraversalPath"
        user_message:
          $ref: "#/components/schemas/MessageSummary"
          description: 저장된 질문 메시지
        web_search_executed:
          type: boolean
          description: 실제로 웹 검색이 수행됐는지. 허용해도 실패하면 false다.
          example: false
        web_search_requested:
          type: boolean
          description: 요청에서 웹 검색을 허용했는지
          example: false
...
    QueryRunStatusResponse:
      type: object
      description: "비동기 질의 상태. 완료 전에는 result가 없고, 실패하면 error가 채워진다."
      properties:
        error:
          type: string
          description: 실패 사유. 성공이면 키가 빠진다.
        model:
          type: string
          description: 이 질의를 처리한 모델명(실행 시점 snapshot)
          example: gpt-5-nano
        provider:
          type: string
          description: 이 질의를 처리한 LLM provider(실행 시점 snapshot)
          example: openai
        request_id:
          type: string
          description: 질의 run ID
          example: query_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
        result:
          $ref: "#/components/schemas/QueryResponse"
          description: 완료된 질의의 결과. 완료 전에는 키가 빠진다.
        status:
          type: string
          description: 현재 상태
          example: completed
        web_search_enabled:
          type: boolean
          description: 이 질의에 웹 검색이 허용됐는지(실행 시점 snapshot)
          example: false
    QuerySourceRef:
      type: object
      description: 근거가 가리키는 문서·block 쌍. 한 근거가 여러 문서를 참조할 수 있어 쌍으로 표현한다.
      properties:
        source_block_id:
          type: string
          description: 문서 안에서 근거가 있는 block ID
        source_document_id:
          type: string
          description: 근거가 있는 문서 ID
          example: doc_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
    RelatedPage:
      type: object
      description: 답변 근거가 된 Wiki 페이지
      properties:
        depth:
          type: integer
          format: int32
          description: 그래프 탐색에서의 거리. 0이면 직접 매칭이다.
          example: 1
        id:
          type: string
          description: Wiki 페이지 ID
        page_type:
          type: string
          description: 페이지 종류
          example: Concept
        relevance_score:
          type: number
          format: double
          description: 질문과의 관련도 점수
          example: 0.87
        role:
          type: string
          description: 이 페이지가 답변에서 맡은 역할
        slug:
          type: string
          description: URL에 쓰는 식별자
          example: search-indexing
        title:
          type: string
          description: 페이지 제목
          example: 검색 인덱싱
...
    TraversalPath:
      type: object
      description: 그래프를 따라간 한 경로. 왜 이 페이지가 근거인지 설명한다.
      properties:
        edges:
          type: array
          description: 경로를 이루는 관계 목록(순서대로)
          items:
            $ref: "#/components/schemas/GraphEdge"
        nodes:
          type: array
          description: 경로가 지나간 페이지 ID 목록(순서대로)
          items:
            type: string
        path_id:
          type: string
          description: 경로 ID
        role:
          type: string
          description: 이 경로가 답변에서 맡은 역할
        score:
          type: number
          format: double
          description: 경로 점수
          example: 0.72
        stop_reason:
          type: string
          description: 탐색을 멈춘 이유
        used_for_answer:
          type: boolean
          description: 이 경로가 실제 답변 생성에 쓰였는지 여부
          example: true
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      name: bearerAuth
      scheme: bearer
      type: http